add day 15 part 2 solution

+++ day15/priorityQueue.go
@@ -5,7 +5,7 @@ import (
 )
 
 type vertexPQ struct {
-	level     byte
+	level     int
 	neighbors []*vertexPQ
 	priority  int
 	index     int
